Экран со списком исключений за текущий день, правки

+++ app/src/main/java/com/example/exceptions/logManager/ExceptionLogManager.kt
@@ -1,14 +1,15 @@
 package com.example.exceptions.logManager
 
 import android.content.Context
-import android.content.Intent
+import android.util.Log
+import com.example.exceptions.logException.DateTimeUtils.toDateTime
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import java.io.File
 import java.util.Calendar
+import java.util.Locale
 
 class ExceptionLogManager(context: Context) {
 
@@ -19,7 +20,7 @@ class ExceptionLogManager(context: Context) {
     data class ExceptionData(
         val thread: String? = null,
         val throwable: String? = null,
-        val timestamp: Long? = null,
+        val timestamp: String? = null,
     )
 
     var enabled = false
@@ -27,36 +28,32 @@ class ExceptionLogManager(context: Context) {
 
     private val gson = Gson()
     private val listType = object : TypeToken<List<ExceptionData>>() {}.type
-    private val calendar = Calendar.getInstance()
     private val countDay = 7
     private val dayInMilliseconds = 86400000L
 
     private val fileManager = FileManager(context)
-    private val currentFileName get() = Calendar.getInstance().apply {
-        set(Calendar.HOUR_OF_DAY, 0)
-        set(Calendar.MINUTE, 0)
-        set(Calendar.SECOND, 0)
-        set(Calendar.MILLISECOND, 0)
-    }.timeInMillis.toString() + ".json"
+    private val currentFileName
+        get() = Calendar.getInstance().apply {
+            set(Calendar.HOUR_OF_DAY, 0)
+            set(Calendar.MINUTE, 0)
+            set(Calendar.SECOND, 0)
+            set(Calendar.MILLISECOND, 0)
+        }.timeInMillis.toString() + ".json"
 
 
     companion object {
         private val _exceptionHandler: MutableStateFlow<ExceptionData?> = MutableStateFlow(null)
         val exceptionHandler: StateFlow<ExceptionData?> get() = _exceptionHandler.asStateFlow()
-
-        const val CUSTOM_INTENT_ACTION = "exceptions.logManager"
-        const val EXCEPTION_KEY = "exceptionKey"
     }
 
     init {
-
         Thread.setDefaultUncaughtExceptionHandler { thread: Thread?, throwable: Throwable? ->
             throwable?.let { throwableObject ->
                 if (enabled) {
                     val newExceptionData = ExceptionData(
                         thread.toString(),
                         throwable.toString(),
-                        calendar.timeInMillis
+                        System.currentTimeMillis().toDateTime()
                     )
                     val listData = convertDataToList(fileManager.readFromFile(currentFileName))
                     listData.add(newExceptionData)
@@ -64,10 +61,6 @@ class ExceptionLogManager(context: Context) {
                     fileManager.writeToFile(outString, currentFileName)
                     lastException = Pair(thread, throwable)
                     if (isShowLog) {
-                        val intent = Intent(CUSTOM_INTENT_ACTION).apply {
-                            putExtra(EXCEPTION_KEY, throwableObject.toString())
-                        }
-                        context.sendBroadcast(intent)
                         _exceptionHandler.value = newExceptionData
                     } else {
                         throwLastExceptionToDefaultHandler()
@@ -77,7 +70,10 @@ class ExceptionLogManager(context: Context) {
         }
 
         fileManager.getFileListFromDirectory().forEach { currentFile ->
-            if (currentFile.toLong() / dayInMilliseconds > countDay) {
+            if (currentFile.lowercase(Locale.getDefault()).endsWith(".json") &&
+                    fileManager.getFileNameWithoutExtension(currentFile).toLong() / dayInMilliseconds > countDay * dayInMilliseconds) {
+                val result = fileManager.getFileNameWithoutExtension(currentFile).toLong() / dayInMilliseconds
+                Log.d("deleteFile", result.toString())
                 fileManager.deleteLocalFile(currentFileName)
             }
         }
